+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/FilmDbStorage.java
@@ -194,7 +194,7 @@ public class FilmDbStorage implements FilmsStorage {
     }
 
     @Override
-    public List<Film> getCommonFilms(long userId, long friendId) {
+    public List<Film> getCommonFilms(int userId, int friendId) {
         log.debug("Common films are: ");
         String query = "SELECT f.* FROM film f " +
                 "WHERE f.film_id IN (SELECT l1.film_id FROM film_likes l1 WHERE l1.user_id = ?) " +
